working/requirements.txt
# Unified Transformer Requirements
# Dependencies for the Unified Transformer model implementation

# Core PyTorch
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data processing
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.5.0

# Scientific computing
scipy>=1.7.0
scikit-learn>=1.0.0

# Configuration and logging
pyyaml>=6.0
tensorboard>=2.10.0
tqdm>=4.62.0

# Image augmentations (optional)
albumentations>=1.1.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Export and deployment (optional)
onnx>=1.12.0
onnxruntime>=1.12.0